# improve performance of RegistrationKey